fot_continued_pretraining/requirements.freeze.txt
-f https://storage.googleapis.com/jax-releases/libtpu_releases.html
--extra-index-url https://download.pytorch.org/whl/cpu
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
chex==0.1.7
click==8.1.7
cloudpickle==2.2.1
clu==0.0.8
colorama==0.4.6
contextlib2==21.6.0
DataProperty==1.0.1
datasets==2.14.2
decorator==5.1.1
dill==0.3.7
distrax==0.1.3
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.6.1
etils==1.3.0
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.3
flatbuffers==23.5.26
flax==0.7.0
frozenlist==1.4.0
fsspec==2022.11.0
gast==0.4.0
gcsfs==2022.11.0
gitdb==4.0.10
GitPython==3.1.34
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
keras==2.11.0
libclang==16.0.6
libtpu-nightly==0.1.dev20230622
lm-eval==0.3.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mbstrdecoder==1.1.3
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.2.0
mlxu==0.1.11
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.22.4
oauthlib==3.2.2
openai==0.28.0
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.2.3
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
pathvalidate==3.1.0
portalocker==2.7.0
protobuf==3.19.6
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycountry==22.3.5
pydantic==2.3.0
pydantic-core==2.6.3
Pygments==2.16.1
pytablewriter==1.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rouge-score==0.1.2
rsa==4.9
sacrebleu==1.5.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlitedict==2.1.0
starlette==0.27.0
sympy==1.12
tabledata==1.3.1
tcolorpy==0.1.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.33.0
tensorflow-probability==0.20.1
tensorstore==0.1.41
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1+cpu
tqdm==4.66.1
tqdm-multiprocess==0.0.11
transformers==4.31.0
typepy==1.3.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.23.2
wandb==0.15.9
werkzeug==2.3.7
wrapt==1.15.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
